Send query to APIs

dbh_json_query returns query in JSON format @description Help function which send query form r to DBH-API, it is designs in the way that it looks like api query and allows the same functionality as in DBH-API

dbh_json_query(tabell_id, filters = list(), group_by = list(),
  sort_by = list(), exclude = NULL)

Arguments

tabell_id
a code name for dataset

filters
is the same as filters in DBH-API: item, all, top, between, greater than

group_by
group by variables in the same way as in DBH-API

sort_by
sort variables

exclude
variable values we do not want

Value

query